AdditionalErrorsDecorator = FallibleOperationsDecorator + InternalServerErrorDecorator
